feat(api): add plunk and refactor nextauth api endpoint config
