+++ package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "@headlessui/react": "^1.7.3",
     "@next-auth/prisma-adapter": "^1.0.5",
+    "@plunk/node": "^0.0.7",
     "@popperjs/core": "^2.11.6",
     "@prisma/client": "^4.7.0",
     "@supabase/supabase-js": "^2.0.5",
