+++ pages/api/auth/[...nextauth].ts
@@ -1,14 +1,6 @@
-import NextAuth, { Session } from "next-auth";
+import NextAuth from "next-auth";
 import AzureADProvider from "next-auth/providers/azure-ad";
-import { PrismaAdapter } from "@next-auth/prisma-adapter";
-
-import prisma from "@/lib/prisma";
-import { ExtendedSession } from "@/types/Session";
-import { InvitationStatus } from "@prisma/client";
-
-const adapter = PrismaAdapter(prisma);
-const oldLinkAccount = adapter.linkAccount;
-adapter.linkAccount = ({ ext_expires_in, ...data }) => oldLinkAccount(data);
+import adapter, { sessionCallback } from "@/lib/nextauth";
 
 export default NextAuth({
   providers: [
@@ -26,32 +18,6 @@ export default NextAuth({
     verifyRequest: "/",
   },
   callbacks: {
-    async session({ session, user }) {
-      // Send properties to the client, like an access_token from a provider.
-      session.user = user;
-
-      const userDetails = await prisma.user.findUnique({
-        where: { id: user.id },
-        include: {
-          preferences: true,
-          receivedInvitations: {
-            include: { event: true, sender: true, receiver: true },
-            where: { status: InvitationStatus.PENDING },
-          },
-          sendedInvitations: true,
-          participations: true,
-          createdEvents: true,
-        },
-      });
-      if (!userDetails) throw new Error("User not found");
-
-      // extend session with user details
-      session = {
-        ...session,
-        ...userDetails,
-      };
-
-      return session as ExtendedSession;
-    },
+    session: sessionCallback,
   },
 });
